Neo-6M GPS module location streaming
It was determined that it may be an easier jumping off point to use a pre-built module to examine some NMEA messages and develop the peripherical modules required for the FPGA-implemented GPS (display/graphics, message passing, process control) rather than developing the FPGA GPS and peripherical functions in parallel

This contains the core interface for a pre-built UBlox Neo-6M communicating via a simple UART connection (as well as the considerable set-up from the vanilla Raspberry Pi configuration). This will eventually be plugged into a stand-in GPS messaging process which will be in C rather than Python. If the UART connection is feasible for the FPGA implementation, maybe it'll be kept and things will be very plug-and-play. Even if not, most of the peripheral functionality will be compartmentalized so this is probably a suitable testbed

testing: Lat/Long coordinates are correct (see below)
contributor: Henry Dikeman

06:39:39 -- (Lat,Long) = (37.4036205,-122.11536033333333)
06:39:40 -- (Lat,Long) = (37.4036205,-122.11536116666667)
06:39:41 -- (Lat,Long) = (37.40362,-122.11536166666667)
06:39:42 -- (Lat,Long) = (37.4036195,-122.11536233333334)
06:39:43 -- (Lat,Long) = (37.40361866666667,-122.11536366666667)

^^ this is actually where I'm at (don't tell anybody)